% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome_install.R
\name{fontawesome_install}
\alias{fontawesome_install}
\title{Check FontAwesome font availability}
\usage{
fontawesome_install()
}
\value{
A message indicating whether FontAwesome is available
}
\description{
\code{fontawesome_install()} checks if FontAwesome is installed on your system
and provides installation instructions if not.
}
\details{
Note: FontAwesome must be installed on your computer.
}
\seealso{
\code{\link{fontawesome_test}}
}
