% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome_test.R
\name{fontawesome_test}
\alias{fontawesome_test}
\title{Test for FontAwesome font availability}
\usage{
fontawesome_test()
}
\value{
Logical \code{TRUE} if FontAwesome is available, \code{FALSE} otherwise
}
\description{
\code{fontawesome_test()} checks if FontAwesome is installed on your system.
FontAwesome is used for glyphs with \code{library(waffle)}.
}
