% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_bar}
\alias{geom_bar}
\title{geom_bar in the Urban Institute style}
\usage{
geom_bar(...)
}
\arguments{
\item{...}{other arguments passed to \code{geom_bar()}}
}
\description{
Submit \code{?ggplot2::geom_line} to see the full documentation for \code{geom_bar()}
}
