% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_col}
\alias{geom_col}
\title{geom_col in the Urban Institute style}
\usage{
geom_col(mapping = NULL, width = 0.7, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{width}{column width}

\item{...}{other arguments passed to \code{geom_col()}}
}
\description{
Submit \code{?ggplot2::geom_line} to see the full documentation for \code{geom_col()}
}
