% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_jitter}
\alias{geom_jitter}
\title{geom_jitter in the Urban Institute style}
\usage{
geom_jitter(mapping = NULL, size = 3, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{point size}

\item{...}{other arguments passed to \code{geom_jitter()}}
}
\description{
Submit \code{?ggplot2::geom_jitter} to see the full documentation for \code{geom_jitter()}
}
