% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_line}
\alias{geom_line}
\title{geom_line in the Urban Institute style}
\usage{
geom_line(mapping = NULL, size = 1, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{line size}

\item{...}{other arguments passed to \code{geom_line()}}
}
\description{
Submit \code{?ggplot2::geom_line} to see the full documentation for \code{geom_line()}
}
