% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_path}
\alias{geom_path}
\title{geom_path in the Urban Institute style}
\usage{
geom_path(mapping = NULL, size = 1, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{line size}

\item{...}{other arguments passed to \code{geom_path()}}
}
\description{
Submit \code{?ggplot2::geom_path} to see the full documentation for \code{geom_path()}
}
