% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_point}
\alias{geom_point}
\title{geom_point in the Urban Institute style}
\usage{
geom_point(mapping = NULL, size = 3, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{point size}

\item{...}{other arguments passed to \code{geom_point()}}
}
\description{
Submit \code{?ggplot2::geom_point} to see the full documentation for \code{geom_point()}
}
