% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_step}
\alias{geom_step}
\title{geom_step in the Urban Institute style}
\usage{
geom_step(mapping = NULL, size = 1, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{line size}

\item{...}{other arguments passed to \code{geom_step()}}
}
\description{
Submit \code{?ggplot2::geom_step} to see the full documentation for \code{geom_step()}
}
