% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_text}
\alias{geom_text}
\title{geom_point in the Urban Institute style}
\usage{
geom_text(mapping = NULL, size = 1/0.352777778, ...)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{size}{text size}

\item{...}{other arguments passed to \code{geom_point()}}
}
\description{
Submit \code{?ggplot2::geom_point} to see the full documentation for \code{geom_point()}
}
