% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{get_legend}
\usage{
get_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{a ggplot}
}
\value{
a ggplot legend
}
\description{
Extract just the legend from a ggplot2 object.
}
