% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_import.R, R/lato_install.R
\name{lato_install}
\alias{lato_install}
\title{Import and register Lato font}
\usage{
lato_install()

lato_install()
}
\description{
\code{lato_import()} tests to see if Lato is imported and registered. If
Lato, isn't imported and registered, then \code{lato_import()} imports and
registers Lato with
\href{https://github.com/wch/extrafont}{library(extrafont)}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{lato_install()} has been deprecated because its name was ambiguous. Please use \code{lato_import()}.
}
\details{
Note: Lato must be installed on your computer for \code{lato_import()} to
work. Lato is the Urban Institute's main font and can be installed from
\href{https://fonts.google.com/specimen/Lato}{Google fonts}.

Test to see if Lato is imported and registered with \code{lato_test()}.
}
