% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_test.R
\name{lato_test}
\alias{lato_test}
\title{Test for Lato font availability}
\usage{
lato_test()
}
\value{
Logical \code{TRUE} if Lato is available, \code{FALSE} otherwise
}
\description{
\code{lato_test()} checks if Lato is installed on your system using
\code{systemfonts}. Lato is the Urban Institute's main font and can be
installed from \href{https://fonts.google.com/specimen/Lato}{Google fonts}.
}
\seealso{
\code{\link{lato_import}} for installation instructions
}
