% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_urbn.R
\name{palette_urbn}
\alias{palette_urbn}
\title{The Urban Institute \link[ggplot2:ggplot2]{ggplot2} theme}
\format{
A \code{list}.
}
\usage{
palette_urbn
}
\description{
Palette data for the Urban Institute
}
\details{
Data used by the palettes in the uithemes package.
}
