% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_legend.R
\name{remove_legend}
\alias{remove_legend}
\title{remove_legend}
\usage{
remove_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{a ggplot}
}
\value{
a ggplot without a legend
}
\description{
Remove just the legend from a ggplot2 object.
}
