% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gradient}
\alias{scale_color_gradient}
\alias{scale_colour_gradient}
\title{Two-color gradient color scale that aligns with the Urban Institute style}
\usage{
scale_color_gradient(
  palette = "cyan",
  low,
  high,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  ...
)

scale_colour_gradient(
  palette = "cyan",
  low,
  high,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  ...
)
}
\arguments{
\item{palette}{Name of the Urban Institute palette to use. Options:
"cyan" (default), "gray", "yellow", "magenta", "green",
"spacegray", "red"}

\item{low}{Color for low end of gradient. If not specified, uses
first color from palette.}

\item{high}{Color for high end of gradient. If not specified, uses
last color from palette.}

\item{space}{colour space in which to calculate gradient.
Must be "Lab" (default) - other values are deprecated.}

\item{na.value}{Color to use for missing values}

\item{guide}{Type of legend. defaults to "colourbar"}

\item{...}{other arguments passed to \code{continuous_scale()}}
}
\description{
Creates a smooth gradient between two colors. By default uses the lightest
and darkest colors from the selected palette.
}
