% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_ordinal}
\alias{scale_color_ordinal}
\title{Discrete color scale for ordinal factors that aligns with the Urban Institute style}
\usage{
scale_color_ordinal(palette = "cyan", ...)
}
\arguments{
\item{palette}{Name of the Urban Institute palette to use. Options:
"cyan" (default), "gray", "yellow", "magenta", "green",
"spacegray", "red", "diverging"}

\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
This function is designed for ordinal/sequential data with up to 8 categories.
Uses sequential palettes (light to dark) by default.
}
