% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urban.R
\name{scale_color_urban}
\alias{scale_color_urban}
\title{Discrete scale color that aligns with the Urban Institute style}
\usage{
scale_color_urban(palette = "categorical", ...)
}
\arguments{
\item{palette}{Urban Institute palette to use. One of \code{categorical} or \code{sequential}}

\item{...}{Additional arguments passed to ggplot2::discrete_scale}
}
\description{
Discrete scale color that aligns with the Urban Institute style
}
