% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Discrete color scale that aligns with the Urban Institute style}
\usage{
scale_colour_discrete(palette = "main", ...)
}
\arguments{
\item{palette}{Name of the Urban Institute palette to use. Options:
"main" (default), "cyan", "gray", "yellow", "magenta", "green",
"spacegray", "red", "diverging", "politics", "quintile"}

\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
This function is designed for up to 8 categories/colors. If more than 8
colors are requested, a warning will be issued and colors will be
interpolated, which may not meet Urban Institute style guidelines.
}
