% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_gradientn}
\alias{scale_colour_gradientn}
\title{Continuous fill scale that aligns with the Urban Institute style}
\usage{
scale_colour_gradientn(
  palette = "cyan",
  ...,
  colours,
  colors,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar"
)
}
\arguments{
\item{palette}{Name of the Urban Institute palette to use for the gradient.
Options: "cyan" (default), "gray", "yellow", "magenta", "green",
"spacegray", "red", "diverging". Ignored if \code{colours} or \code{colors}
is specified.}

\item{...}{other arguments passed to \code{discrete_scale()}}

\item{colours}{vector of colours}

\item{colors}{vector of colours}

\item{values}{if colours should not be evenly positioned along the gradient this vector gives the position (between 0 and 1) for each colour in the colours vector. See rescale for a convience function to map an arbitrary range to between 0 and 1}

\item{space}{colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.}

\item{na.value}{default color for NA values}

\item{guide}{legend representation of scale}
}
\description{
Continuous fill scale that aligns with the Urban Institute style
}
