% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urban.R
\name{scale_colour_urban}
\alias{scale_colour_urban}
\title{Discrete scale color that aligns with the Urban Institute style}
\usage{
scale_colour_urban(palette = "categorical", ...)
}
\arguments{
\item{palette}{Urban Institute palette to use. One of \code{categorical} or \code{sequential}}
}
\description{
Discrete scale color that aligns with the Urban Institute style
}
