% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urbn.R
\name{scale_fill_urbn}
\alias{scale_fill_urbn}
\title{Discrete scale fill that aligns with the Urban Institute style}
\usage{
scale_fill_urbn(palette = "categorical", ...)
}
\arguments{
\item{palette}{Urban Institute palette to use. One of \code{categorical} or \code{sequential}}
}
\description{
Discrete scale fill that aligns with the Urban Institute style
}
