% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_urban_default}
\alias{set_urban_default}
\title{Set default color for monochromatic plots.}
\usage{
set_urban_default(color = "#1696d2")
}
\arguments{
\item{color}{Color to be set. Defaults to Urban Institute blue #1696d2.}
}
\description{
Update a set of geoms and stats for monochromatic plots. Defaults to Urban
Institute blue #1696d2.
}
