% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_urban_defaults.R
\name{set_urban_defaults}
\alias{set_urban_defaults}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
set_urban_defaults()
}
\description{
\code{set_urban_defaults} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide for web, with sensible defaults.
}
