% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-urban.R
\name{theme_urban}
\alias{theme_urban}
\title{A \link{ggplot2} theme formatted in the Urban Institute style}
\usage{
theme_urban(base_family = "Lato", base_size = 12, plot_title_size = 18L,
  plot_title_face = "bold", plot_title_margin = 8,
  plot_subtitle_size = 14L, plot_subtitle_face = "plain",
  plot_subtitle_margin = 10, plot_caption_size = 8L,
  plot_caption_face = "plain", plot_caption_margin = base_size/2 * 0.9,
  plot_margin = margin(10L, 10L, 10L, 10L), axis_title_size = 12L,
  axis_title_face = "italic", axis_text_size = 12L,
  axis_text_face = "plain", legend_position = "top",
  legend_direction = "horizontal", legend_box = "horizontal",
  panel_ontop = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_size, plot_title_face, plot_title_margin}{plot title size typeface and margin}

\item{plot_subtitle_size, plot_subtitle_face, plot_subtitle_margin}{plot subtitle size typeface and margin}

\item{plot_caption_size, plot_caption_face, plot_caption_margin}{plot caption size typeface and margin}

\item{plot_margin}{plot margin (specify with \link[ggplot2:margin]{ggplot2::margin})}

\item{axis_title_size, axis_title_face}{axis title size and typeface}

\item{axis_text_size, axis_text_face}{axis title size and typeface}

\item{legend_position, legend_direction, legend_box}{legend position layout and arrangement}

\item{panel_ontop}{place the panel over the data layers}
}
\description{
\code{theme_urban} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide, with sensible defaults.
}
