% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_urbn_web.R
\name{theme_urbn_web}
\alias{theme_urbn_web}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
theme_urbn_web(base_size = 12L, base_family = "Lato",
  base_line_size = base_size/24L, base_rect_size = base_size/24L)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_urbn_web} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide for web, with sensible defaults.
}
