% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_urban_defaults.R
\name{undo_urban_defaults}
\alias{undo_urban_defaults}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
undo_urban_defaults()
}
\description{
Resets all the \link{ggplot2} theme to its default settings
}
