% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urban.R
\name{urban_color_pal}
\alias{urban_color_pal}
\title{Color Palettes based on Urban Institute theme}
\usage{
urban_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used in the Urban Institute.
}
