% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_geofacet.R
\docType{data}
\name{urbn_geofacet}
\alias{urbn_geofacet}
\title{Urban Institute geofacet template}
\format{Data frame with columns
\describe{
\item{row}{Row in geofacet}
\item{col}{Column in geofacet}
\item{state_abbv}{State abbreviation}
\item{state_name}{State name}
}}
\usage{
urbn_geofacet
}
\description{
Urban Institute geofacet template
}
\keyword{datasets}
