% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_logo_text.R
\name{urbn_logo_text}
\alias{urbn_logo_text}
\title{The Urban Institute \link[ggplot2:ggplot2]{ggplot2} theme}
\usage{
urbn_logo_text()
}
\description{
Creates a grid object with the Urban Institute logo
}
