% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{palette_urbn_main}
\alias{palette_urbn_main}
\alias{palette_urbn_diverging}
\alias{palette_urbn_quintile}
\alias{palette_urbn_politics}
\alias{palette_urbn_cyan}
\alias{palette_urbn_gray}
\alias{palette_urbn_yellow}
\alias{palette_urbn_magenta}
\alias{palette_urbn_green}
\alias{palette_urbn_spacegray}
\alias{palette_urbn_red}
\title{Urban Institute palettes}
\source{
Urban Institute Data Visualization Style Guide
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\usage{
palette_urbn_main

palette_urbn_diverging

palette_urbn_quintile

palette_urbn_politics

palette_urbn_cyan

palette_urbn_gray

palette_urbn_yellow

palette_urbn_magenta

palette_urbn_green

palette_urbn_spacegray

palette_urbn_red
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Urban Institute Data Visualization Style Guide.
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\concept{urbn palettes}
