% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_plot.R
\name{urbn_plot}
\alias{urbn_plot}
\title{urbn_plot}
\usage{
urbn_plot(..., ncol = 1, heights = 1)
}
\arguments{
\item{...}{urbn plot objects or grobs}

\item{ncol}{number of columns to arrange objects into}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
Combine elements from \code{urbn_title}, \code{urbn_subtitle},
\code{urbn_y_title}, \code{get_legend}, \code{remove_legend},
\code{urbn_notes}, \code{urbn_source}, and \code{ggplot2} into
one formatted plot.
}
\details{
This function requires the \code{patchwork} package.
}
