% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_source.R
\name{urbn_source}
\alias{urbn_source}
\title{urbn_source}
\usage{
urbn_source(string, size = 8)
}
\arguments{
\item{string}{character string for a source statement}

\item{size}{font size for the source}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
urbn_source
}
