% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_subtitle.R
\name{urbn_subtitle}
\alias{urbn_subtitle}
\title{urbn_subtitle}
\usage{
urbn_subtitle(string, size = 9.5)
}
\arguments{
\item{string}{character string for a subtitle}

\item{size}{font size for the subtitle}
}
\value{
a grob formatted for a subtitle in a ggplot
}
\description{
Add a subtitle to a plot created with \code{urbn_plot()}.
}
