% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_title.R
\name{urbn_title}
\alias{urbn_title}
\title{urbn_title}
\usage{
urbn_title(string, size = 12, width = 68)
}
\arguments{
\item{string}{character string for a title}

\item{size}{font size for the title}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
urbn_title
}
