% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_y_title.R
\name{urbn_y_title}
\alias{urbn_y_title}
\title{urbn_y_title}
\usage{
urbn_y_title(string, size = 8.5)
}
\arguments{
\item{string}{character string for a y-axis title}

\item{size}{font size for the y-axis title}
}
\value{
a grob formatted for a y-axis title in a ggplot
}
\description{
Add a horizontal y axis title to a plot created with \code{urbn_plot()}.
}
