% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axis.R
\name{add_axis}
\alias{add_axis}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
add_axis(axis = "y")
}
\arguments{
\item{axis}{Axis to drop. "x", "y", or "both"}
}
\description{
Adds tick axes and axes labels
}
