% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome_install.R
\name{fontawesome_install}
\alias{fontawesome_install}
\title{Import and register Lato font}
\usage{
fontawesome_install()
}
\description{
\code{fontawesome_install} tests to see if FontAwesome is imported and registered. If FontAwesome, isn't imported and registered, then \code{fontawesome_install} imports and registers FontAwesome with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: FontAwesome must be installed on your computer for \code{fontawesome_install} to work.

Test to see if FontAwesome is imported and registered with \code{fontawesome_test()}.
}
