% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome_test.R
\name{fontawesome_test}
\alias{fontawesome_test}
\title{Test for FontAwesome import and registration}
\usage{
fontawesome_test()
}
\description{
\code{fontawesome_test} tests to see if FontAwesome is imported and registered. FontAwesome is used for glyphs with \code{library(waffle)} and can be installed from \href{https://github.com/hrbrmstr/waffle/tree/master/inst/fonts}{Bob Rudis's GitHub}.
}
\details{
Import and register FontAwesome in R with fontawesome_install().
}
