% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_install.R
\name{lato_install}
\alias{lato_install}
\title{Import and register Lato font}
\usage{
lato_install()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{lato_install()} was deprecated because its name was ambiguous. Please use \code{lato_import()}.
}
