% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_install.R
\name{lato_install}
\alias{lato_install}
\title{Import and register Lato font}
\usage{
lato_install()
}
\description{
\code{lato_install} tests to see if Lato is imported and registered. If Lato, isn't imported and registered, then \code{lato_install} imports and registers Lato with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Lato must be installed on your computer for \code{lato_install} to work. Lato is the Urban Institute's main font and can be installed from \href{https://fonts.google.com/specimen/Lato}{Google fonts}.

Test to see if Lato is imported and registered with \code{lato_test()}.
}
