% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_test.R
\name{lato_test}
\alias{lato_test}
\title{A \link{ggplot2} theme formatted in the Urban Institute style}
\usage{
lato_test()
}
\description{
\code{lato_test} tests to see if Lato is imported and registered. Lato is the Urban Institute's main font and can be installed from \href{https://fonts.google.com/specimen/Lato}{Google fonts}.
}
\details{
Import and register Lato in R with lato_install().
}
