% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-urban.R
\docType{data}
\name{palette_urban}
\alias{palette_urban}
\title{Palette data for the Urban Institute}
\format{A \code{list}.}
\usage{
palette_urban
}
\description{
Data used by the palettes in the uithemes package.
}
\keyword{datasets}
