% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_urbn.R
\docType{data}
\name{palette_urbn}
\alias{palette_urbn}
\title{The Urban Institute [ggplot2] theme}
\format{A \code{list}.}
\usage{
palette_urbn
}
\description{
Palette data for the Urban Institute
}
\details{
Data used by the palettes in the uithemes package.
}
\keyword{datasets}
