% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axis.R, R/remove_axis.R
\name{remove_axis}
\alias{remove_axis}
\alias{remove_axis}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
remove_axis(axis = "y")

remove_axis(axis = "y")
}
\arguments{
\item{axis}{Axis to drop. "x", "y", or "both"}

\item{axis}{Axis to drop. "x", "y", or "both"}
}
\description{
Adds tick axes and axes labels

Removes axes and axes labels
}
