% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ticks.R
\name{remove_ticks}
\alias{remove_ticks}
\title{The Urban Institute \link{ggplot2} theme}
\usage{
remove_ticks(axis = "both")
}
\arguments{
\item{axis}{Axis to drop. "x", "y", or "both"}
}
\description{
Removes tick marks from all axes
}
