% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gradientn}
\alias{scale_color_gradientn}
\title{Continuous scale fill that aligns with the Urban Institute style}
\usage{
scale_color_gradientn(...)
}
\description{
Continuous scale fill that aligns with the Urban Institute style
}
