% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urban.R
\name{scale_color_gradientn_urban}
\alias{scale_color_gradientn_urban}
\title{Continuous scale fill that aligns with the Urban Institute style}
\usage{
scale_color_gradientn_urban(..., colours = c("#CFE8F3", "#A2D4EC", "#73BFE2",
  "#46ABDB", "#1696D2", "#12719E", "#0A4C6A", "#062635"),
  colors = c("#CFE8F3", "#A2D4EC", "#73BFE2", "#46ABDB", "#1696D2", "#12719E",
  "#0A4C6A", "#062635"), values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar")
}
\description{
Continuous scale fill that aligns with the Urban Institute style
}
