% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Discrete scale color that aligns with the Urban Institute style}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{palette}{Urban Institute palette to use. One of \code{categorical} or \code{sequential}}
}
\description{
Discrete scale color that aligns with the Urban Institute style
}
