% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Discrete color scale that aligns with the Urban Institute style}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
Discrete color scale that aligns with the Urban Institute style
}
