% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_ordinal}
\alias{scale_colour_ordinal}
\title{Discrete color scale for ordinal factors that aligns with the Urban Institute style}
\usage{
scale_colour_ordinal(...)
}
\description{
Discrete color scale for ordinal factors that aligns with the Urban Institute style
}
