% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Discrete fill scale that aligns with the Urban Institute style}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
Discrete fill scale that aligns with the Urban Institute style
}
