% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_ordinal}
\alias{scale_fill_ordinal}
\title{Discrete fill scale for ordinal factors that aligns with the Urban Institute style}
\usage{
scale_fill_ordinal(...)
}
\description{
Discrete fill scale for ordinal factors that aligns with the Urban Institute style
}
