% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_urbn_print.R
\name{theme_urbn_print}
\alias{theme_urbn_print}
\title{A \link{ggplot2} theme formatted in the Urban Institute style}
\usage{
theme_urbn_print(
  base_size = 8.5,
  base_family = "Lato",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_urban} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide for web, with sensible defaults.
}
