% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_urbn_defaults.R
\name{undo_urbn_defaults}
\alias{undo_urbn_defaults}
\title{The Urban Institute \link[ggplot2:ggplot2]{ggplot2} theme}
\usage{
undo_urbn_defaults()
}
\description{
Resets all the \link[ggplot2:ggplot2]{ggplot2} theme to its default settings
}
