% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-urban.R, R/scales.R
\name{urban_color_pal}
\alias{urban_color_pal}
\alias{urban_color_pal}
\title{The Urban Institute [ggplot2] theme}
\usage{
urban_color_pal(palette = "categorical")

urban_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}

\item{palette}{Palette name.}
}
\description{
Color palettes used in the Urban Institute.

Color palettes used in the Urban Institute.
}
