% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{urbn_color_pal}
\alias{urbn_color_pal}
\title{The Urban Institute [ggplot2] theme}
\usage{
urbn_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used in the Urban Institute.
}
