% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_geofacet.R
\docType{data}
\name{urbn_geofacet}
\alias{urbn_geofacet}
\title{Dataset used for creating geofacetted plots in the Urban Institute style.}
\format{
Data frame with columns
\describe{
\item{row}{Row in geofacet}
\item{col}{Column in geofacet}
\item{code}{State abbreviation}
\item{name}{State name}
}
}
\usage{
urbn_geofacet
}
\description{
Dataset used for creating geofacetted plots in the Urban Institute style.
}
\keyword{datasets}
