% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbn_plot.R
\name{urbn_plot}
\alias{urbn_plot}
\title{urbn_plot}
\usage{
urbn_plot(..., heights)
}
\arguments{
\item{...}{urbn plot objects or grobs}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
urbn_plot
}
