% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{The Urban Institute \link[ggplot2:ggplot2]{ggplot2} theme}
\usage{
view_palette(palette = palette_urbn_main)
}
\arguments{
\item{palette}{A \code{palette_urbn_*} vector from \code{library(urbnthemes)}.
Options are \code{palette_urbn_diverging}, \code{palette_urbn_politics},
\code{palette_urbn_quintile}, \code{palette_urbn_cyan}, \code{palette_urbn_gray},
\code{palette_urbn_yellow}, \code{palette_urbn_magenta}, \code{palette_urbn_green},
\code{palette_urbn_spacegray}, and \code{palette_urbn_red}.}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for \code{palette_urbn_*} vectors.
}
\examples{
view_palette()
view_palette(palette_urbn_cyan)

}
